feature: add group access rule and dedicated user list
This change introduces a rule where one may specify user groups inside
access control blocks as well as define a dedicated list of users that are
members of these groups. A typical user entry consists of a username,
an auth key and a number of group memberships. A matching auth key in
combination with the correct group membership grants access under this rule.