Implement basic JWT auth rule
This version of JwtAuthSyncRule checks whether a request has
an Authorization header containing a bearer token.

If it does and it can be determined that it is a correctly signed,
claim based JWT then the rule matches. Otherwise, NO_MATCH is
returned.

The key used to validate the token's signature is configured by
the rule settings under the key "signature.key". For instance:

    jwt_auth:
      signature.key: "a secret"

Currently, it only supports signature algorithms where the
key can be used as is (sans a String::getBytes call), like
those based on HMAC (HS256, etc).