Implement 'jwt_auth.user_claim' setting
Once a token's signature is verified, claims embedded
in the payload can be used to populate variables
that subsequent rules may rely on.

One of those variables is '@user'. It represents the
logged in user identification.

A new setting, keyed 'jwt_auth.user_claim', is introduced
to let the rule know which of the claims present in the token
it should use to call 'RequestContext::setLoggedInUser' with.