jwt_auth rule: authenticate users through JSON Web Tokens (#218)
* Depend on io.jsonwebtoken:jjwt:0.7.0

* Implement basic JWT auth rule

This version of JwtAuthSyncRule checks whether a request has
an Authorization header containing a bearer token.

If it does and it can be determined that it is a correctly signed,
claim based JWT then the rule matches. Otherwise, NO_MATCH is
returned.

The key used to validate the token's signature is configured by
the rule settings under the key "signature.key". For instance:

    jwt_auth:
      signature.key: "a secret"

Currently, it only supports signature algorithms where the
key can be used as is (sans a String::getBytes call), like
those based on HMAC (HS256, etc).

* Implement 'jwt_auth.user_claim' setting

Once a token's signature is verified, claims embedded
in the payload can be used to populate variables
that subsequent rules may rely on.

One of those variables is '@user'. It represents the
logged in user identification.

A new setting, keyed 'jwt_auth.user_claim', is introduced
to let the rule know which of the claims present in the token
it should use to call 'RequestContext::setLoggedInUser' with.

* Add integration tests for the 'jwt_auth' rule

* Use gradle generated licenses

* Add creator for 'jwt_auth¡ to the settings creators registry

* Use a string as signature.key value

* Wrap token parsing in doPrivileged call

* Add 'reject-expired-token' integration test

* Automated integration tests for jwt_auth

* Support 'env:' and 'text:' prefixes

The signature key configuration supports the mentioned
prefixes with the following meaning:

* env:<var> will lookup <var> in the environment and
  use the value there as the key
* text:<key> will use <key> as the key, handy if your
  plaintext key starts with 'env:'

If no prefix is found, the configuration is used as if
it were prefixed by 'text:'.

* Handle some corner cases in JwtAuthRuleSettings

* Empty keys are not allowed
* Key must be a YAML string

If any of those conditions are not met, a
SettingsMalformedException is throwed.

* Rename setting 'signature.key' to 'signature_key'